requirements.txt
# requirements.txt  
pytesseract>=0.3.10  
paddleocr>=2.7.0  
opencv-python>=4.6.0,<4.7.0  
Pillow>=9.0.0  
mlflow>=2.0.0  
numpy<2  